Refactor models into separate data package
The models from models.go were split into separate files (user.go, consultant.go, etc.) within a new data package. The migration function in db.go was updated accordingly to import and use these models from the new package instead. This refactoring enhances the code organization and readability.
