+++ data/user.go
@@ -1,0 +1,11 @@
+package data
+
+import (
+	"gorm.io/gorm"
+)
+
+type User struct {
+	gorm.Model
+	Username       string
+	HashedPassword string
+}
