+++ data/workprofile.go
@@ -1,0 +1,11 @@
+package data
+
+import (
+	"gorm.io/gorm"
+)
+
+type WorkProfile struct {
+	gorm.Model
+	ConsultantRefer uint
+	Consultant      Consultant `gorm:"foreignKey:ConsultantRefer;constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
+}
